WIP - refactor Transaction.value() from 'Double' to 'BigDecimal' + add BigDecimalSummaryStatistics + add BigDecimalSummaryStatisticsTest

+++ src/main/java/com/example/itau_backend/dto/inbound/TransactionRequest.java
@@ -1,10 +1,11 @@
 package com.example.itau_backend.dto.inbound;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 import jakarta.validation.constraints.NotNull;
 
 public record TransactionRequest(
-    @NotNull Double valor,
+    @NotNull BigDecimal valor,
     @NotNull OffsetDateTime dataHora) {
 }
